refactor(sub-agent-executor): modularize sub-agent-executor.js (1417 LOC)
Extract lib/sub-agent-executor.js into 11 focused modules:
- constants.js (43 LOC) - Configuration values and thresholds
- phase-model-config.js (128 LOC) - Model routing configuration
- supabase-client.js (26 LOC) - Supabase client singleton
- sd-resolver.js (118 LOC) - SD key/UUID resolution
- model-routing.js (33 LOC) - Phase-aware model selection
- pattern-loader.js (57 LOC) - Pattern injection
- instruction-loader.js (145 LOC) - Instruction loading
- executor.js (370 LOC) - Main execution logic
- results-storage.js (284 LOC) - Database storage
- history.js (130 LOC) - History queries
- index.js (169 LOC) - Re-exports

All modules under 500 LOC. Original file becomes re-export wrapper.
DESIGN validation: 100% confidence.

Part of SD-LEO-REFACTOR-LARGE-FILES-003 (child 9/10)
