+++ lib/sub-agent-executor/instruction-loader.js
@@ -1,0 +1,145 @@
+/**
+ * Instruction Loader
+ * Loads and formats sub-agent instructions from database
+ *
+ * Extracted from sub-agent-executor.js for modularity
+ * SD-LEO-REFACTOR-SUBAGENT-EXEC-001
+ */
+
+import { getSupabaseClient } from './supabase-client.js';
+import { loadRelevantPatterns } from './pattern-loader.js';
+
+/**
+ * Load sub-agent instructions from database
+ * @param {string} code - Sub-agent code (e.g., 'VALIDATION', 'TESTING', 'DATABASE')
+ * @returns {Promise<Object>} Sub-agent data with formatted instructions
+ */
+export async function loadSubAgentInstructions(code) {
+  console.log(`\nLoading sub-agent instructions: ${code}...`);
+
+  const supabase = await getSupabaseClient();
+
+  const { data: subAgent, error } = await supabase
+    .from('leo_sub_agents')
+    .select('*')
+    .eq('code', code)
+    .single();
+
+  if (error) {
+    throw new Error(`Failed to load sub-agent ${code} from database: ${error.message}`);
+  }
+
+  if (!subAgent) {
+    throw new Error(`Sub-agent ${code} not found in database`);
+  }
+
+  // LEO Protocol v4.3.2 Enhancement: Load relevant patterns
+  const relevantPatterns = await loadRelevantPatterns(code);
+
+  // Format instructions for Claude to read (now includes patterns)
+  const formatted = formatInstructionsForClaude(subAgent, relevantPatterns);
+
+  console.log(`Loaded: ${subAgent.name} (v${subAgent.metadata?.version || '1.0.0'})`);
+
+  return {
+    ...subAgent,
+    relevantPatterns,
+    formatted
+  };
+}
+
+/**
+ * Format sub-agent instructions for Claude to read
+ * LEO Protocol v4.3.2: Now includes relevant patterns and prevention checklists
+ * @param {Object} subAgent - Sub-agent record from database
+ * @param {Array} relevantPatterns - Relevant issue patterns (optional)
+ * @returns {string} Formatted instructions
+ */
+export function formatInstructionsForClaude(subAgent, relevantPatterns = []) {
+  const metadata = subAgent.metadata || {};
+  const capabilities = subAgent.capabilities || [];
+  const sources = metadata.sources || [];
+
+  // Format pattern section
+  let patternSection = '';
+  if (relevantPatterns && relevantPatterns.length > 0) {
+    patternSection = `
+────────────────────────────────────────────────────────────────
+KNOWN ISSUES & PROVEN SOLUTIONS (from issue_patterns)
+────────────────────────────────────────────────────────────────
+`;
+    relevantPatterns.forEach((p, i) => {
+      const severityIcon = p.severity === 'critical' ? '[CRITICAL]' : p.severity === 'high' ? '[HIGH]' : '[MEDIUM]';
+      const topSolution = p.proven_solutions && p.proven_solutions.length > 0
+        ? p.proven_solutions[0].solution || p.proven_solutions[0].method || 'See pattern details'
+        : 'No proven solution yet';
+
+      patternSection += `
+${i + 1}. ${severityIcon} [${p.pattern_id}] ${p.issue_summary}
+   Category: ${p.category} | Occurrences: ${p.occurrence_count} | Trend: ${p.trend}
+   Proven Solution: ${topSolution.substring(0, 100)}${topSolution.length > 100 ? '...' : ''}
+`;
+    });
+
+    // Add aggregated prevention checklist
+    const preventionItems = new Set();
+    relevantPatterns.forEach(p => {
+      if (p.prevention_checklist) {
+        p.prevention_checklist.slice(0, 2).forEach(item => preventionItems.add(item));
+      }
+    });
+
+    if (preventionItems.size > 0) {
+      patternSection += `
+────────────────────────────────────────────────────────────────
+PREVENTION CHECKLIST (Apply Before Proceeding)
+────────────────────────────────────────────────────────────────
+`;
+      Array.from(preventionItems).slice(0, 5).forEach((item, i) => {
+        patternSection += `[ ] ${i + 1}. ${item}\n`;
+      });
+    }
+  }
+
+  return `
+════════════════════════════════════════════════════════════════
+${subAgent.name} (${subAgent.code})
+Version: ${metadata.version || '1.0.0'}
+Priority: ${subAgent.priority || 50}
+════════════════════════════════════════════════════════════════
+
+${subAgent.description || 'No description available'}
+
+${capabilities.length > 0 ? `
+────────────────────────────────────────────────────────────────
+CAPABILITIES
+────────────────────────────────────────────────────────────────
+${capabilities.map((c, i) => `${i + 1}. ${c}`).join('\n')}
+` : ''}
+
+${sources.length > 0 ? `
+────────────────────────────────────────────────────────────────
+LESSONS SOURCES
+────────────────────────────────────────────────────────────────
+${sources.map((s, i) => `${i + 1}. ${s}`).join('\n')}
+` : ''}
+${patternSection}
+${metadata.success_patterns ? `
+────────────────────────────────────────────────────────────────
+SUCCESS PATTERNS
+────────────────────────────────────────────────────────────────
+${metadata.success_patterns.map((p, i) => `${i + 1}. ${p}`).join('\n')}
+` : ''}
+
+${metadata.failure_patterns ? `
+────────────────────────────────────────────────────────────────
+FAILURE PATTERNS TO AVOID
+────────────────────────────────────────────────────────────────
+${metadata.failure_patterns.map((p, i) => `${i + 1}. ${p}`).join('\n')}
+` : ''}
+
+════════════════════════════════════════════════════════════════
+END OF INSTRUCTIONS
+════════════════════════════════════════════════════════════════
+`;
+}
